# Conflicts:
#	src/main/java/swp/auctionkoi/exception/ErrorCode.java
#	src/main/java/swp/auctionkoi/models/Auction.java
#	src/main/java/swp/auctionkoi/repository/AuctionRequestRepository.java
#	src/main/java/swp/auctionkoi/service/auction/impl/AuctionServiceImpl.java
#	src/main/java/swp/auctionkoi/service/user/impl/StaffServiceImpl.java

+++ src/main/java/swp/auctionkoi/models/Auction.java
@@ -25,7 +25,7 @@ public class Auction {
     swp.auctionkoi.models.KoiFish fish;
 
     @Column(name = "current_price")
-    float currentPrice;
+    Double currentPrice;
 
     @Column(name = "status")
     Integer status;
